Add script for comparing two JSON files
We are going to change the repository structure, build tools and
tests. Since we will need to make commits that change both the build
tool and tests, we will need a way to check manually that the generated
files do not change between commits. For that, we need to be able to
compare two JSON files while ignoring the order of items in arrays. The
order does not matter in our data structures, and it can differ between
implementations. The `compare.py` script does just that.