Remove BATS-based tests
A previous commit has already disabled the original BATS-based tests for
the build.sh script. We have eventually agreed to get rid of the
build.sh script and have the service [1] call whatever new build tools
directly. This allows us to remove the BATS-based tests and reduce the
number of technologies used in the repository.

Note: The commit preserves the build.sh script with no tests for now
because it is still responsible for determining the version of
configuration data being built. This functionality will be moved to
the Python build script(s); the `build.sh` script will be removed then.

[1] https://github.com/RedHatInsights/insights-operator-gathering-conditions-service/blob/main/get_conditions.sh