Rework build.py and add tests for happy day scenarios
WARNING: BREAKING CHANGE

This is a breaking change with respect to the Insights Operator
Gathering Conditions Service.

We originally intended to preserve the interface that the
get_conditions.sh script in the service repo expected of the build tool
in this repository. However, it has proven problematic from the
testability and documentation point of view.

The build tool is easier to document if it produces a directory
structure that copies the structure of source data. Additional changes
have been done to improve testability of the build tool.

Note that the build tool still does not implement all validation and
error handling that is planned. It will be added in separate commits
later.