Fix build tool test 'invalid_corrupted_v1_config_schema'
The purpose of the test is to check the build tool behavior in case v1
remote config schema is corrupted. For the test to be robust (with
respect to build tool refactoring), the schemas should be complete and
valid except for `remote_configuration_v1.schema.json` (which uses
`BadType` as a property type).