Change terminology from "templates" to "blueprints"
For each remote configuration file required by an Insights Operator in
the fleet, the source data contain a file that specifies which
conditional gathering rules and container log requests the remote
configuration file should contain. We used to call these files
"templates", which suggested incorrectly that each "template" is used to
generate multiple different remote configuration files.

The term "blueprint" seems to match better the purpose and usage of
those files.

Disclaimer: No one on the team is a native English speaker so the
terminology might still be misleading. :-)