Use relative paths in cluster_version_mapping.json
The `cluster_version_mapping.json` file used to reference remote
configurations (or their blueprints) by filename, expecting they would
be in the `remote_configurations` subdirectory of the
`cluster_version_mapping.json` parent directory. This was awkward and
difficult to explain, as the previous sentence demonstrates.

This commit changes the `cluster_version_mapping.json` file to reference
remote configurations (or their blueprints) using paths relative to the
location of the file.